+++ src/plugin/components/VInlineCustomField/VInlineCustomField.vue
@@ -174,7 +174,7 @@ watch(() => loadingProp.value, (newVal, oldVal) => {
 // ------------------------------------------------ The displayed value //
 const displayValue = computed(() => {
 	if (modelValue.value) {
-		empty.value = false;
+		setEmptyValue(false);
 
 		if (settings.truncateLength) {
 			return useTruncateText({
@@ -187,9 +187,13 @@ const displayValue = computed(() => {
 		return modelValue.value;
 	}
 
-	empty.value = true;
+	setEmptyValue(true);
 	return settings.emptyText;
 });
+
+function setEmptyValue(val: boolean) {
+	empty.value = val;
+}
 
 
 // ------------------------------------------------ Binding Events & Props //